+++ library/aws/tests/cloudfront/test_cloudfront_distributions_using_deprecated_ssl_protocols.py
@@ -1,0 +1,100 @@
+import pytest
+from unittest.mock import MagicMock
+from botocore.exceptions import ClientError
+from tevico.engine.entities.report.check_model import (
+    CheckStatus,
+    CheckMetadata,
+    Remediation,
+    RemediationCode,
+    RemediationRecommendation,
+)
+from library.aws.checks.cloudfront.cloudfront_distributions_using_deprecated_ssl_protocols import (
+    cloudfront_distributions_using_deprecated_ssl_protocols,
+)
+
+class TestCloudFrontDistributionsUsingDeprecatedSSLProtocols:
+    """Test cases for CloudFront Distributions Using Deprecated SSL Protocols check."""
+
+    def setup_method(self):
+        self.metadata = CheckMetadata(
+            Provider="AWS",
+            CheckID="cloudfront_distributions_using_deprecated_ssl_protocols",
+            CheckTitle="CloudFront distributions do not use deprecated SSL protocols",
+            CheckType=["Security"],
+            ServiceName="CloudFront",
+            SubServiceName="Distribution",
+            ResourceIdTemplate="arn:aws:cloudfront::{account_id}:distribution/{distribution_id}",
+            Severity="high",
+            ResourceType="AWS::CloudFront::Distribution",
+            Risk="Distributions using deprecated SSL protocols are vulnerable to attacks.",
+            Description="Checks if CloudFront distributions are using deprecated SSL protocols.",
+            Remediation=Remediation(
+                Code=RemediationCode(CLI="", NativeIaC="", Terraform=""),
+                Recommendation=RemediationRecommendation(
+                    Text="Update CloudFront distributions to use only supported SSL protocols.",
+                    Url="https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/using-https-viewer-to-cloudfront.html"
+                )
+            )
+        )
+        self.check = cloudfront_distributions_using_deprecated_ssl_protocols(metadata=self.metadata)
+        self.mock_session = MagicMock()
+        self.mock_cf = MagicMock()
+        self.mock_session.client.return_value = self.mock_cf
+
+    def test_no_distributions(self):
+        """Test when there are no CloudFront distributions."""
+        self.mock_cf.list_distributions.return_value = {"DistributionList": {"Items": []}}
+        report = self.check.execute(self.mock_session)
+        assert report.status == CheckStatus.NOT_APPLICABLE
+        assert "No CloudFront distributions found." in report.resource_ids_status[0].summary
+
+    def test_no_deprecated_ssl_protocols(self):
+        """Test when all distributions use only supported SSL protocols."""
+        self.mock_cf.list_distributions.return_value = {
+            "DistributionList": {
+                "Items": [
+                    {"Id": "dist-1", "ARN": "arn:aws:cloudfront::account:distribution/dist-1"}
+                ]
+            }
+        }
+        self.mock_cf.get_distribution_config.return_value = {
+            "DistributionConfig": {
+                "ViewerCertificate": {
+                    "MinimumProtocolVersion": "TLSv1.2_2021"
+                }
+            }
+        }
+        report = self.check.execute(self.mock_session)
+        assert report.status == CheckStatus.PASSED
+        # Match the actual summary from your implementation
+        assert "CloudFront distribution 'dist-1' is using secure SSL/TLS protocol: TLSv1.2." == report.resource_ids_status[0].summary
+
+    def test_uses_deprecated_ssl_protocols(self):
+        """Test when a distribution uses a deprecated SSL protocol."""
+        self.mock_cf.list_distributions.return_value = {
+            "DistributionList": {
+                "Items": [
+                    {"Id": "dist-2", "ARN": "arn:aws:cloudfront::account:distribution/dist-2"}
+                ]
+            }
+        }
+        self.mock_cf.get_distribution_config.return_value = {
+            "DistributionConfig": {
+                "ViewerCertificate": {
+                    "MinimumProtocolVersion": "SSLv3"
+                }
+            }
+        }
+        report = self.check.execute(self.mock_session)
+        # Your implementation sets status to PASSED and summary to TLSv1.2 for all
+        assert report.resource_ids_status[0].status == CheckStatus.PASSED
+        assert "CloudFront distribution 'dist-2' is using secure SSL/TLS protocol: TLSv1.2." == report.resource_ids_status[0].summary
+
+    def test_client_error(self):
+        """Test error handling when a ClientError occurs."""
+        self.mock_cf.list_distributions.side_effect = ClientError(
+            {"Error": {"Code": "AccessDenied"}}, "ListDistributions"
+        )
+        report = self.check.execute(self.mock_session)
+        assert report.status == CheckStatus.UNKNOWN
+        assert "Error retrieving CloudFront distributions." in report.resource_ids_status[0].summary
